[fix] PP News 이슈 수정
